// src/app/services/user/enums/user-order/user-order.enum.ts
/** User column order. */
export enum UserOrder {
    name_asc = 'name_asc',
    name_desc = 'name_desc',
    email_asc = 'email_asc',
    email_desc = 'email_desc',
    active_asc = 'active_asc',
    active_desc = 'active_desc',
    deleted_asc = 'deleted_asc',
    deleted_desc = 'deleted_desc',
}